Move color mode button to sidebar menu.
Split app and team Pinia stores.
