+++ components/lineup/LineupJerseyButton.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
-import { useAppSettingsStore } from '~~/stores/AppSettings'
+import { useTeamStore } from '~~/stores/Team'
 
 import type { Spot } from '~~/types';
 
-const appSettingsStore = useAppSettingsStore();
+const teamStore = useTeamStore();
 
 const props = defineProps<{
     spot: Spot,
@@ -16,11 +16,11 @@ const isDialogVisible = ref(false);
 <template>
     <div @click="isDialogVisible = true">
         <svg viewBox="215 45 526.2 372.045" :width="props.size" :height="props.size" xmlns="http://www.w3.org/2000/svg">
-            <use x="-308.27" href="@/assets/images/cloth-t-shirt.svg#shirt-back" :fill="`#${ appSettingsStore.jerseyColor }`" />
+            <use x="-308.27" href="@/assets/images/cloth-t-shirt.svg#shirt-back" :fill="`#${ teamStore.jerseyColor }`" />
 
-            <text v-if="props.spot.player.number.length === 1" x="445" y="200" font-size="10em" font-weight="bold" :fill="`#${ appSettingsStore.jerseyTextColor }`">{{ props.spot.player.number }}</text>
-            <text v-else-if="props.spot.player.number.length === 2" x="390" y="200" font-size="10em" font-weight="bold" :fill="`#${ appSettingsStore.jerseyTextColor }`">{{ props.spot.player.number }}</text>
-            <text v-else-if="props.spot.player.number.length === 3" x="370" y="200" font-size="8.5em" font-weight="bold" :fill="`#${ appSettingsStore.jerseyTextColor }`">{{ props.spot.player.number }}</text>
+            <text v-if="props.spot.player.number.length === 1" x="445" y="200" font-size="10em" font-weight="bold" :fill="`#${ teamStore.jerseyTextColor }`">{{ props.spot.player.number }}</text>
+            <text v-else-if="props.spot.player.number.length === 2" x="390" y="200" font-size="10em" font-weight="bold" :fill="`#${ teamStore.jerseyTextColor }`">{{ props.spot.player.number }}</text>
+            <text v-else-if="props.spot.player.number.length === 3" x="370" y="200" font-size="8.5em" font-weight="bold" :fill="`#${ teamStore.jerseyTextColor }`">{{ props.spot.player.number }}</text>
         </svg>
 
         <Dialog
