+++ components/settings/SettingsButton.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import { useAppSettingsStore } from '~~/stores/AppSettings';
+import { useTeamStore } from '~~/stores/Team';
 
-const appSettingsStore = useAppSettingsStore();
+const teamStore = useTeamStore();
 
 const buttonPassThroughOptions = {
   label: {
@@ -28,7 +28,7 @@ const isDialogVisible = ref(false);
   <Dialog
       v-model:visible="isDialogVisible"
       modal
-      header="App Settings"
+      header="Team Settings"
       :pt="{
           root: { class: 'w-full md:w-3/4 xl:w-1/2' },
           content: { class: 'pt-1' }
@@ -36,44 +36,17 @@ const isDialogVisible = ref(false);
     >
     <div class="grid grid-cols-2 md:grid-cols-4 gap-2">
       <div class="text-right text-xs py-2">
-        Color Mode
-      </div>
-      <div class="md:col-span-3">
-        <Button
-            @click="appSettingsStore.colorMode = appSettingsStore.colorMode === 'system' ? 'dark' : appSettingsStore.colorMode === 'dark' ? 'light' : 'system'"
-            size="small"
-            severity="primary"
-            rounded
-            aria-label="Color mode"
-            title="Color mode"
-          >
-          <template v-if="appSettingsStore.colorMode === 'system'">
-            <i class="pi pi-fw pi-desktop pr-2"></i>
-            System
-          </template>
-          <template v-if="appSettingsStore.colorMode === 'dark'">
-            <i class="pi pi-fw pi-moon pr-2"></i>
-            Dark
-          </template>
-          <template v-if="appSettingsStore.colorMode === 'light'">
-            <i class="pi pi-fw pi-sun pr-2"></i>
-            Light
-          </template>
-        </Button>
-      </div>
-
-      <div class="text-right text-xs py-2">
         Jersey Color
       </div>
       <div>
-        <ColorPicker v-model="appSettingsStore.jerseyColor" format="hex" />
+        <ColorPicker v-model="teamStore.jerseyColor" format="hex" />
       </div>
 
       <div class="text-right text-xs py-2">
         Jersey Text Color
       </div>
       <div>
-        <ColorPicker v-model="appSettingsStore.jerseyTextColor" format="hex" />
+        <ColorPicker v-model="teamStore.jerseyTextColor" format="hex" />
       </div>
     </div>
   </Dialog>
