+++ stores/AppSettings.ts
@@ -1,7 +1,4 @@
-import { defineStore, skipHydrate, acceptHMRUpdate } from "pinia";
-import { nanoid } from "nanoid";
-
-import type { Spot, Lineup } from '~~/types';
+import { defineStore, acceptHMRUpdate } from "pinia";
 
 // NOTE: This Pinia store requires `ssr: false` in `nuxt.config.*` because the server
 //       will try to serialize our computed state properties to JSON, and that will fail
@@ -16,19 +13,6 @@ export const useAppSettingsStore = defineStore('AppSettingsStore', {
   state: () => {
     const _colorMode = useColorMode();
 
-    const defaults = {
-      id: nanoid(),
-
-      teamName: '',
-      jerseyColor: 'f47373',
-      jerseyTextColor: '000000',
-
-      isLocked: false,
-      spots: [],
-    };
-
-    const _lineup = useLocalStorage<Lineup>('AppSettingsStore:lineup', defaults, { mergeDefaults: true });
-
     // NOTE: We provide setters so we can persist to storage.
     //       Source: https://github.com/vuejs/pinia/issues/447#issuecomment-1455285437
 
@@ -37,63 +21,16 @@ export const useAppSettingsStore = defineStore('AppSettingsStore', {
       set: (v) => _colorMode.preference = v
     });
 
-    const teamName = computed<string>({
-      get: () => _lineup.value.teamName,
-      set: (v) => _lineup.value.teamName = v
-    });
-
-    const jerseyColor = computed<string>({
-      get: () => _lineup.value.jerseyColor,
-      set: (v) => _lineup.value.jerseyColor = v
-    });
-
-    const jerseyTextColor = computed<string>({
-      get: () => _lineup.value.jerseyTextColor,
-      set: (v) => _lineup.value.jerseyTextColor = v
-    });
-
-    const isLocked = computed<boolean>({
-      get: () => _lineup.value.isLocked,
-      set: (v) => _lineup.value.isLocked = v
-    });
-
-    const spots = computed<Spot[]>({
-      get: () => _lineup.value.spots,
-      set: (v) => _lineup.value.spots = v
-    });
-
     // function $reset() {
     //   //TODO
     // }
 
     return {
-      //TODO we may need to skipHydrate() if we ever reenable SSR
-
       colorMode,
-
-      teamName,
-      jerseyColor,
-      jerseyTextColor,
-
-      isLocked,
-      spots,
 
       //$reset,
     };
   },
-  actions: {
-    addSpot(spot: Spot) {
-      if (!this.isLocked) {
-        this.spots.push({ ...spot });
-      }
-    },
-    removeSpot(playerId: string) {
-      this.spots = this.spots.filter(s => s.player.id !== playerId)
-    },
-  },
-  // getters: {
-
-  // }
 });
 
 if (import.meta.hot) {
